refactor and move to utils
